PR-E-4: extract SlotPlanner (slot-allocator collectors)
Move the 16 BoundTreeWalker slot-allocator collector classes out of the
monolithic ReflectionMetadataEmitter.cs into a dedicated
'internal sealed class SlotPlanner' at src/Core/CodeAnalysis/Emit/SlotPlanner.cs.

Collectors moved (all nested 'private sealed' inside SlotPlanner):
- PatternSwitchSlotAllocator
- StructLiteralCollector
- AppendCollector
- NullConditionalCollector
- BlockExpressionLocalCollector
- ExpressionBlockLabelCollector
- LambdaCollector
- GoStatementCollector
- GoCapturedVariableCollector
- DefaultExpressionCollector
- ReceiverSpillCollector
- MapIndexReadCollector
- IndexAssignmentValueSpillCollector
- AssignmentValueSpillCollector
- NullableValueTypeUnwrapCollector
- NullableValueTypeCoalesceCollector

Also moved: 'SelectSlots' value-object (promoted to top-level
'internal sealed class' in the same file so existing call sites need no
namespace prefix change).

Deliberately NOT moved: ConstructedTypeCollector. It is a
BoundTreeRewriter (not a walker), tied to the closure/display-class
synthesis path. It belongs with PR-E-9 ClosureEmitter, where the rest
of the closure emit machinery moves.

Architecture (Option A): the per-method-emit slot dictionaries
(structLiteralSlots, mapIndexSlots, defaultExpressionSlots,
receiverSpillSlots, indexAssignmentValueSlots, appendSlots, etc.)
continue to live where they did (in CollectLocalsAndLabels and on the
BodyEmitter nested class), and are passed into each collector at the
call site. SlotPlanner has no static state and depends only on
EmitContext, MetadataTokenCache, and a 'needs spill?' Func<...>
predicate bound to the existing ReflectionMetadataEmitter.NeedsRvalueReceiverSpill.
This keeps the diff narrow and preserves per-method ownership of
slot state, which is the right shape for a future invariant that
'every BoundNode that needs a slot has one before any reader runs'
(coming as a post-decomposition follow-up).

Wiring: ReflectionMetadataEmitter constructs 'new SlotPlanner(emitCtx, cache, NeedsRvalueReceiverSpill)'
once and exposes public entry-point methods on SlotPlanner
(CollectStructLiterals, CollectAppends, CollectMapIndexReads,
CollectIndexAssignmentValueSpills, CollectDefaultExpressions,
CollectReceiverSpills, CollectAssignmentValueSpills,
CollectNullableValueTypeUnwraps, CollectNullableValueTypeCoalesces,
CollectBlockExpressionLocals, CollectExpressionBlockLabels,
CollectFunctionLiterals, CollectGoStatements, CollectCapturedVariables,
CollectNullConditionalCaptures, WalkForPatternSwitches,
WalkExpressionForSwitches, WalkPatternForSwitchExpressions,
WalkForPatternSwitchAllocation). The existing wrapper methods on
ReflectionMetadataEmitter now delegate to those.

SlotDictionaryAliasingAssertionTests: the asserted defensive substrings
(!structLiteralSlots.ContainsKey(literal), !mapIndexSlots.ContainsKey(idx),
!defaultExpressionSlots.ContainsKey(def),
'if (receiverSpillSlots.ContainsKey(receiver))',
'if (receiverSpillSlots.ContainsKey(assn))') remain in
ReflectionMetadataEmitter.cs at the allocation sites in
CollectLocalsAndLabels. The test globs every src/Core/CodeAnalysis/Emit/*.cs
file, so it passes regardless of which sibling file housed each substring.

Invariants preserved:
- Public/internal API surface: unchanged. No renames visible to
  Compilation or tests.
- IL byte-identical: RefactoringBaselineTests shows zero hash diffs.
- DeterministicEmitTests: green.
- All GS#### diagnostics: same codes, locations, messages.

LoC: ReflectionMetadataEmitter.cs 17217 -> 16284 (-933);
SlotPlanner.cs new at 1085.

Next in sequence: PR-E-5 ConversionEmitter.
